[core] Merge annotation_tile_data.{hpp,cpp} into annotation_tile.{hpp,cpp}
