+++ src/mbgl/annotation/annotation_tile.hpp
@@ -1,8 +1,28 @@
 #pragma once
 
 #include <mbgl/tile/geometry_tile.hpp>
+#include <mbgl/tile/geometry_tile_data.hpp>
 
 namespace mbgl {
+
+class AnnotationManager;
+
+namespace style {
+class UpdateParameters;
+}
+
+class AnnotationTileData : public GeometryTileData {
+public:
+    AnnotationTileData(const OverscaledTileID&,
+                       std::string sourceID,
+                       const style::UpdateParameters&);
+    ~AnnotationTileData();
+
+    void setNecessity(Necessity) final;
+
+private:
+    AnnotationManager& annotationManager;
+};
 
 class AnnotationTileFeature : public GeometryTileFeature {
 public:
