+++ src/mbgl/style/source.cpp
@@ -20,9 +20,9 @@
 #include <mbgl/util/tile_cover.hpp>
 
 #include <mbgl/tile/raster_tile_data.hpp>
-#include <mbgl/tile/annotation_tile_data.hpp>
 #include <mbgl/tile/geojson_tile.hpp>
 #include <mbgl/tile/vector_tile_data.hpp>
+#include <mbgl/annotation/annotation_tile.hpp>
 #include <mbgl/style/parser.hpp>
 #include <mbgl/gl/debugging.hpp>
